.NET 8 executable version
This is a release build for .NET 8 for a late beta of DIAMetric.  It doesn't have finalized CV terms for its mzQC output, and it doesn't yet feature the ability to read DIA-NN outputs (a feature that I just haven't committed the time to create).  I haven't updated the GUI code to use .NET 8, so it is not included here.